WIIS-63: mise en place système de rôles dynamiques + menus + actions (reste encore reliquats rôles en dur -> à suivre)

+++ src/Controller/FournisseurController.php
@@ -60,11 +60,7 @@ class FournisseurController extends AbstractController
      */
     public function index(): Response
     {
-        if ($this->isGranted('ROLE_ADMIN_GT')) {
-            return $this->render('fournisseur/index.html.twig', ['fournisseur' => $this->fournisseurRepository->findAll()]);
-        } else {
-            return new Response($this->renderView('securite/access_denied.html.twig'));
-        }
+        return $this->render('fournisseur/index.html.twig', ['fournisseur' => $this->fournisseurRepository->findAll()]);
     }
 
 
@@ -107,16 +103,20 @@ class FournisseurController extends AbstractController
      */
     public function edit(Request $request): Response
     {
-        if (!$request->isXmlHttpRequest() && $data = json_decode($request->getContent(), true)) {
-            $fournisseur = $this->fournisseurRepository->find($data['id']);
-            $fournisseur
-                ->setNom($data['nom'])
-                ->setCodeReference($data['CodeReference']);
-            $em = $this->getDoctrine()->getManager();
-            $em->flush();
-            return new JsonResponse();
+        if ($this->isGranted('ROLE_ADMIN_GT')) {
+            if (!$request->isXmlHttpRequest() && $data = json_decode($request->getContent(), true)) {
+                $fournisseur = $this->fournisseurRepository->find($data['id']);
+                $fournisseur
+                    ->setNom($data['nom'])
+                    ->setCodeReference($data['CodeReference']);
+                $em = $this->getDoctrine()->getManager();
+                $em->flush();
+                return new JsonResponse();
+            }
+            throw new NotFoundHttpException("404");
+        } else {
+            //TODO CG message erreur (pas les droits pour cette action)
         }
-        throw new NotFoundHttpException("404");
     }
 
     /**
@@ -124,14 +124,18 @@ class FournisseurController extends AbstractController
      */
     public function delete(Request $request): Response
     {
-        if (!$request->isXmlHttpRequest() && $data = json_decode($request->getContent(), true)) {
-            $fournisseur = $this->fournisseurRepository->find($data['fournisseur']);
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->remove($fournisseur);
-            $entityManager->flush();
-            return new JsonResponse();
+        if ($this->isGranted('ROLE_ADMIN_GT')) {
+            if (!$request->isXmlHttpRequest() && $data = json_decode($request->getContent(), true)) {
+                $fournisseur = $this->fournisseurRepository->find($data['fournisseur']);
+                $entityManager = $this->getDoctrine()->getManager();
+                $entityManager->remove($fournisseur);
+                $entityManager->flush();
+                return new JsonResponse();
+            }
+            throw new NotFoundHttpException("404");
+        } else {
+            //TODO CG message erreur (pas les droits pour cette action)
         }
-        throw new NotFoundHttpException("404");
     }
 
     /**
